+++ packages/patrol_cli/lib/src/crossplatform/flutter_tool.dart
@@ -59,11 +59,14 @@ class FlutterTool {
     }
 
     Future<void> onQuitWithRevertInteractiveMode() async {
-      if (previousStdinModes != null) {
-        revertInteractiveMode(previousStdinModes);
-      }
-      if (onQuit != null) {
-        await onQuit();
+      try {
+        if (previousStdinModes != null) {
+          revertInteractiveMode(previousStdinModes);
+        }
+      } finally {
+        if (onQuit != null) {
+          await onQuit();
+        }
       }
     }
 
@@ -139,7 +142,7 @@ class FlutterTool {
                 '${dartDefine.key}=${dartDefine.value}',
               ],
             ])
-            ..disposedBy(scope);
+            ..disposedByTree(scope);
 
       final completer = Completer<void>();
       scope.addDispose(() {
@@ -179,7 +182,6 @@ class FlutterTool {
               _logger.success(helpText.toString());
             } else if (char == 'q' || char == 'Q') {
               _logger.success('Quitting process...');
-              process.kill();
               if (!completer.isCompleted) {
                 completer.complete();
               }
@@ -269,7 +271,7 @@ class FlutterTool {
               '--device-id',
               deviceId,
             ], runInShell: true)
-            ..disposedBy(scope);
+            ..disposedByTree(scope);
 
       final completer = Completer<void>();
       scope.addDispose(() {
@@ -342,8 +344,9 @@ class FlutterTool {
   }
 
   void revertInteractiveMode(StdinModes stdinModes) {
-    io.stdin.echoMode = stdinModes.echoMode;
+    // Windows allows setting echoMode only while lineMode is enabled.
     io.stdin.lineMode = stdinModes.lineMode;
+    io.stdin.echoMode = stdinModes.echoMode;
 
     _logger.detail('Interactive shell mode disabled.');
   }