+++ packages/patrol_cli/lib/src/macos/macos_test_backend.dart
@@ -120,7 +120,7 @@ class MacOSTestBackend {
               runInShell: true,
               workingDirectory: _rootDirectory.childDirectory('macos').path,
             )
-            ..disposedBy(scope);
+            ..disposedByTree(scope);
       process.listenStdOut((l) => _logger.detail('\t$l')).disposedBy(scope);
       process.listenStdErr((l) => _logger.err('\t$l')).disposedBy(scope);
       exitCode = await process.exitCode;
@@ -177,7 +177,7 @@ class MacOSTestBackend {
               },
               workingDirectory: _rootDirectory.childDirectory('macos').path,
             )
-            ..disposedBy(_disposeScope);
+            ..disposedByTree(_disposeScope);
       process.listenStdOut((l) => _logger.detail('\t$l')).disposedBy(scope);
       process.listenStdErr((l) => _logger.err('\t$l')).disposedBy(scope);